backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.24.0
pydantic[email]>=2.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
google-genai
langchain==0.3.13
langchain-community>=0.2.0
langchain-core>=0.2.0
chromadb>=0.5.0
sentence-transformers>=2.2.2
numpy>=1.22.4,<2.0.0
supabase==2.16.0
aiohttp==3.9.1